+++ jopa-api/src/main/java/cz/cvut/kbss/jopa/model/descriptors/ObjectPropertyCollectionDescriptor.java
@@ -19,12 +19,14 @@ package cz.cvut.kbss.jopa.model.descriptors;
 
 import cz.cvut.kbss.jopa.model.metamodel.FieldSpecification;
 
+import java.lang.reflect.Field;
 import java.net.URI;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
 /**
- * Allows to provide descriptor for elements of an object property collection.
+ * Allows providing descriptor for elements of an object property collection.
  */
 public class ObjectPropertyCollectionDescriptor extends FieldDescriptor {
 
@@ -49,6 +51,20 @@ public class ObjectPropertyCollectionDescriptor extends FieldDescriptor {
                                               boolean assertionsInSubjectContext) {
         super(context, attribute);
         this.elementDescriptor = new EntityDescriptor(context, assertionsInSubjectContext);
+    }
+
+    ObjectPropertyCollectionDescriptor(FieldSpecification<?, ?> attribute, EntityDescriptor elementDescriptor) {
+        super(attribute);
+        assert elementDescriptor != null;
+        this.elementDescriptor = elementDescriptor;
+    }
+
+    protected ObjectPropertyCollectionDescriptor(Set<URI> contexts, boolean assertionsInSubjectContext, String language,
+                                                 boolean hasLanguage, boolean includeInferred, Field field,
+                                                 EntityDescriptor elementDescriptor) {
+        super(contexts, assertionsInSubjectContext, language, hasLanguage, includeInferred, field);
+        assert elementDescriptor != null;
+        this.elementDescriptor = elementDescriptor;
     }
 
     @Override
@@ -83,12 +99,30 @@ public class ObjectPropertyCollectionDescriptor extends FieldDescriptor {
     }
 
     @Override
+    public Descriptor addContext(URI context) {
+        if (context == null) {
+            elementDescriptor.contexts.clear();
+            contexts.clear();
+        } else {
+            elementDescriptor.addContext(context);
+            contexts.add(context);
+        }
+        return this;
+    }
+
+    @Override
     public boolean overridesAssertionContext() {
         return false;
     }
 
-    public EntityDescriptor getElementDescriptor() {
+    @Override
+    public EntityDescriptor unwrap() {
         return elementDescriptor;
+    }
+
+    @Override
+    public ObjectPropertyCollectionDescriptor copy() {
+        return new ObjectPropertyCollectionDescriptor(contexts, assertionsInSubjectContext, getLanguage(), hasLanguage(), includeInferred(), getField(), elementDescriptor.copy());
     }
 
     @Override
@@ -100,16 +134,12 @@ public class ObjectPropertyCollectionDescriptor extends FieldDescriptor {
     }
 
     @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (!super.equals(obj))
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        ObjectPropertyCollectionDescriptor other = (ObjectPropertyCollectionDescriptor) obj;
-        if (elementDescriptor == null) {
-            return other.elementDescriptor == null;
-        } else return elementDescriptor.equals(other.elementDescriptor);
+    protected boolean equals(Object other, Map<VisitedPair, Boolean> visited) {
+        return super.equalsImpl(other) && elementDescriptor.equals(((ObjectPropertyCollectionDescriptor) other).elementDescriptor, visited);
+    }
+
+    @Override
+    protected int hashCode(Map<Object, Boolean> visited) {
+        return 31 * super.hashCodeImpl() + elementDescriptor.hashCode();
     }
 }